# openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: zzITEDPPatientEngagementIntegration@boehringer-ingelheim.com
    name: Patient Engagement Integration
    url: https://api-gw-dev.boehringer-ingelheim.com:8443
  description: "Patient Engagement Enterprise API allows multiple communication channels\
    \ (e.g. websites,  HUbs, etc.) to enrroll and/or manage patient's support porgrams\
    \ and services offered by Boehringer Ingelheim brands.\r\n"
  license:
    name: Boehringer Ingelheim intellectual property. All Rights Reserved.  You may
      not use without express written permission from Boehringer Ingelheim.
    url: https://www.boehringer-ingelheim.com
  title: Patient Engagement Enterprise API
  version: 0.7.1
servers:
- description: DEV
  url: https://api-gw-dev.boehringer-ingelheim.com:8443/patient-ent-engagement
- description: QA
  url: https://api-gw-qa.boehringer-ingelheim.com/patient-ent-engagement
- description: PROD
  url: https://api-gw.boehringer-ingelheim.com/patient-ent-engagement
...
paths:
...
  /patients/{id}/consents:
    parameters:
    - description: BI's Unique Patient Identifier
      explode: false
      in: path
      name: id
      required: true
      schema:
        type: string
      style: simple
    patch:
      description: Allows to update patient's consents and preferences for a brand.
        It will also allow to opt-out a patient from all BI Corporate communications
        (which will unsubscribe the patient from all BI's brands and services).
      operationId: patch-patient-consents-brand
      parameters:
      - description: BI's Unique Patient Identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            examples:
              patch-patient-consents:
                value:
                  consentsAndPreferences:
                  - name: cyltezoPatientNavigator
                    status: optIn
                  - name: cyltezoCopayTextMessageCommunications
                    status: optOut
                  dataProviderId: CMMH
                  dataProviderTransactionId: "2398329334"
                  dataProviderPatientId: "10000045678"
              bi-opt-out:
                value:
                  consentsAndPreferences:
                  - name: biCorporateCommunications
                    status: optOut
                  dataProviderId: epsilon
                  dataProviderTransactionId: 03404-39-04390349
                  dataProviderPatientId: "92372309"
            schema:
              $ref: '#/components/schemas/patch_patient_consents_brand_request'
        description: Patient updated consents and preferences
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/patch_patient_consents_brand_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/xml:
              schema:
                properties: {}
                type: object
            application/json:
              schema:
                properties: {}
                type: object
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
      security:
      - oauth2DEV: []
      - oauth2QA: []
      - oauth2PROD: []
      summary: Update Consents and Preferences
      tags:
      - Consent
...
  /patients/{id}/services/{brand}:
    parameters:
    - description: Patient Identifier (MDMID)
      explode: false
      in: path
      name: id
      required: true
      schema:
        type: string
      style: simple
    - description: Brand where the patient has previously enrolled
      explode: false
      in: path
      name: brand
      required: true
      schema:
        type: string
      style: simple
    patch:
      description: Allows to provide Patient Support Program updates relevant to the
        patient services and health information.
      operationId: patch-patients-id-services-brand
      parameters:
      - description: Patient Identifier (MDMID)
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Brand where the patient has previously enrolled
        explode: false
        in: path
        name: brand
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            examples:
              ExampleZ:
                value:
                  dataProviderId: CMMH
                  dataProviderPatientId: CMM83475324850
                  dataProviderTransactionId: "959587457643569846984"
                  clinicalDetails:
                    prescription:
                      drugName: zongertinib
                      drugGroup: TKI
                      medicationTier: tier one
                  insuranceTypes:
                  - commertial
                  insurancePlans:
                  - planType: pharmacy
                    planName: familyhealth01
                    rxBin: "123456"
                    rxPcn: "12345678910"
                    groupNumber: OCJ3048
                    memberId: "379847593745"
                    provider: CVS
                    providerPhoneNumber: 1-800-872-3862
                    policyholderName: Josh Baker
                    policyholderBirthDate: 1984-02-28
                    insuranceHierarchy: primary
                    coinsurancePercentage: 80 percent
                  serviceSubscriptions:
                  - serviceType: caseManagement
                    subscriptionStatus: optIn
                    caseId: CMMCASEKEY123048
                    patientJourneyStatus: 3rd Party Restriction
                    benefitsVerificationStatus: Benefit Investigation completed
                    priorAuthorizationStatus: required
                    priorAuthorizationRequired: "yes"
                    priorAuthorizationExpiredDate: 2025-08-24T14:15:22Z
                    priorAuthorizationId: "8320204"
                  - serviceType: rxBridge
                    subscriptionStatus: optIn
                    subscriptionDate: 2019-08-24
                    bridgeStatus: Shipped
                    triageDate: 2025-08-24T14:15:22Z
                    prescription:
                      drugName: zongertinib
                      status: string
                      shippedDate: 2019-08-24
                      specialtyPharmacyName: CMM Pharma P
                      specialtyPharmacyTriageDate: 2025-08-24T14:15:22Z
                  - serviceType: copay
                    subscriptionDate: 2023-11-13
                    subscriptionStatus: registrationCompleted
                    rxMemberId: NEWZ83100101048
            schema:
              $ref: '#/components/schemas/patch_patients_id_services_brand_request'
      responses:
        "202":
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
      summary: Update Patient Services
      tags:
      - PaCE
components:
  schemas:
    Error:
      anyOf:
      - $ref: '#/components/schemas/ErrorDefault'
      - $ref: '#/components/schemas/ErrorDefined'
      description: Describes that the system encoutered  an error while processing
        the request.
      title: Error
    ErrorDefined:
      description: Describes that the system encoutered  an error while processing
        the request.
      properties:
        status:
          description: the HTTP status code applicable to this problem
          title: status
          type: number
        id:
          description: error unique identifier
          title: id
          type: string
        title:
          description: "A short, human-readable summary of the problem"
          title: title
          type: string
        detail:
          description: explanation specific to this occurrence of the problem
          title: detail
          type: string
        errors:
          description: Schema validation errors of the request
          items:
            type: string
          title: errors
          type: array
      required:
      - id
      - status
      - title
      title: Error Defined
      x-examples:
        example-400:
          status: 400
          id: BI-PATENG-00016
          title: Patient already registered for Copay.
          detail: The request payload provided by the client contains a service subscription
            to Copay services for a patient that is already register for Copay. Please
            correct the payload and try again.
        example-400-Schema:
          success: false
          code: 400
          request-id: 5f9a3e48-5a92-4e98-ad31-aed876227ff4
          message: Failed to validate body
          errors:
          - Failed to validate "application/json" body
          - ".serviceSubscriptions should NOT have duplicate items (items ## 0 and\
            \ 1 are identical)"
    ErrorDefault:
      description: Describes that the system encoutered  an error while processing
        the request.
      example:
        code: 5.962133916683182
        reasons:
        - ""
        - ""
        details:
        - ""
        - ""
        message: message
        request-id: request-id
        errors:
        - errors
        - errors
      properties:
        code:
          description: the HTTP status code applicable to this problem
          title: code
          type: number
        message:
          description: "A short, human-readable summary of the problem"
          title: message
          type: string
        request-id:
          description: Unique transaction request identifier
          title: request-id
          type: string
        errors:
          description: Schema validation errors of the request
          items:
            type: string
          title: errors
          type: array
        reasons:
          items: {}
          title: reasons
          type: array
        details:
          items: {}
          title: details
          type: array
      required:
      - code
      title: Error Default
      x-examples:
        example-400-schema-val:
          success: false
          code: 400
          request-id: 78453a9a-b973-4ff6-9ef0-e5c563c160e7
          message: Failed to validate body
          errors:
          - Failed to validate "application/json" body
          - should have required property 'enrollmentDate'
    DataProviderInfo:
      description: 'Describes the entity or BI Partner that provides the data for
        processing. '
      properties:
        dataProviderId:
          description: "BI assigned partner unique identifier for a particular partner\
            \ channel used to gather the Patient's information. For example if the\
            \ same parter has web and phone changel, then there will be two unique\
            \ identifier for the same partner one for web and another for phone. "
          maxLength: 50
          minLength: 1
          title: dataProviderId
          type: string
        dataProviderPatientId:
          description: 'External BI partner/vendor patient Identifier. If the client
            application does not keep a patient Id (i.e. brand websites) then provide
            the same dataProviderTransactionId. '
          maxLength: 255
          minLength: 1
          title: dataProviderPatientId
          type: string
        dataProviderTransactionId:
          description: "External BI partner transaction Id (for Incident management,\
            \ allows E2E tracing of this request)"
          maxLength: 255
          minLength: 1
          title: dataProviderTransactionId
          type: string
      required:
      - dataProviderId
      - dataProviderPatientId
      - dataProviderTransactionId
      title: DataProvider
...
    Consent:
      description: "Given assent, approval, certification or attestation."
      example:
        preferences:
        - name: bestContactTime
          selectedOptions:
          - morning
          - morning
          - morning
          - morning
          - morning
        - name: bestContactTime
          selectedOptions:
          - morning
          - morning
          - morning
          - morning
          - morning
        - name: bestContactTime
          selectedOptions:
          - morning
          - morning
          - morning
          - morning
          - morning
        - name: bestContactTime
          selectedOptions:
          - morning
          - morning
          - morning
          - morning
          - morning
        - name: bestContactTime
          selectedOptions:
          - morning
          - morning
          - morning
          - morning
          - morning
        name: Patient Support Program Communications
        status: optOut
      properties:
        name:
          description: Unique and common name of the agreement
          example: Patient Support Program Communications
          maxLength: 100
          minLength: 1
          title: name
          type: string
        status:
          description: "Indicates if the patient (or caregiver) gave or not his/her\
            \ permission.\r\n- optIn: agreed to the term and conditions\r\n- optOut:\
            \ a previously given consent (optIn) has been removed.  Only applies after\
            \ an \"optIn\" was provided.\r\n- expired (read ONLY): after providing\
            \ cossent and over a period of time the consnet is not valid any more.\r\
            \n- notGiven (read ONLY): the consent was not provided (optIn). "
          enum:
          - optOut
          - optIn
          - expired
          - notGiven
          title: status
          type: string
        preferences:
          description: 'List of selected options related to the consent. Usually aplies
            to communications options. '
          items:
            $ref: '#/components/schemas/Preference'
          maxItems: 20
          minItems: 0
          title: preferences
          type: array
          uniqueItems: true
      required:
      - name
      - status
      title: Consent
    Preference:
      description: Describes the choices of an Individual related to an specific questions.
      example:
        name: bestContactTime
        selectedOptions:
        - morning
        - morning
        - morning
        - morning
        - morning
      properties:
        name:
          description: Unique identifier of preference
          example: bestContactTime
          maxLength: 100
          minLength: 1
          title: name
          type: string
        selectedOptions:
          description: Preference selected value
          items:
            example: morning
            maxLength: 100
            minLength: 1
            type: string
          maxItems: 10
          minItems: 1
          title: selectedOptions
          type: array
          uniqueItems: true
      required:
      - name
      - selectedOptions
      title: Preference
...
    ClinicalDetails:
      description: Details of the patient's medical condition treated with the BI
        brand
      example:
        priorTreatments:
        - priorTreatments
        - priorTreatments
        - priorTreatments
        - priorTreatments
        - priorTreatments
        drugAllergies:
        - drugAllergies
        - drugAllergies
        - drugAllergies
        - drugAllergies
        - drugAllergies
        brandAdherence: brandAdherence
        prescription:
          quantity: 6027.456183070403
          drugGroup: drugGroup
          addOns: addOns
          drugForm: drugForm
          medicationTier: medicationTier
          refills: 1.4658129805029452
          specialtyPharmacyTriageDate: 2000-01-23T04:56:07.000+00:00
          drugStrength: drugStrength
          signatureDate: 2000-01-23
          specialtyPharmacyName: specialtyPharmacyName
          drugName: drugName
          dispenseAsWritten: dispenseAsWritten
          shippedDate: 2000-01-23
          status: status
          shipTo: patient
        diagnosisDate: 2000-01-23
        supportPersonsNumber: 8.008281904610115
        icdCode: icdCode
        concomitantMedications: concomitantMedications
        indication: indication
        lastDoseDate: 2000-01-23
        therapyStartDate: 2000-01-23
      properties:
        icdCode:
          description: International Classification of Diseases (ICD) code such as
            ICD-10 as published by the World Health Organization.
          maxLength: 20
          minLength: 0
          title: icdCode
          type: string
        indication:
          description: 'Reason to use the medication. '
          maxLength: 255
          minLength: 0
          title: indication
          type: string
        diagnosisDate:
          description: Date when the underlying medical condition was found.
          format: date
          title: diagnosisDate
          type: string
        priorTreatments:
          description: Treatements already tried  to manage the medical indication.
          items:
            maxLength: 255
            minLength: 1
            type: string
          maxItems: 50
          minItems: 1
          title: priorTreatments
          type: array
        concomitantMedications:
          description: 'Drugs used or given at or almost at the same time as part
            of the therapy. '
          maxLength: 255
          minLength: 0
          title: concomitantMedications
          type: string
        brandAdherence:
          description: "Indicates if the patient has taken the brand prescription\
            \ before, or it's the first time. For example: firstTime or takenBefore"
          maxLength: 255
          minLength: 0
          title: brandAdherence
          type: string
        lastDoseDate:
          description: "If taken the drug before, indicates the last time it was administered. "
          format: date
          title: lastDoseDate
          type: string
        therapyStartDate:
          description: Date of BI medication intake
          format: date
          title: therapyStartDate
          type: string
        supportPersonsNumber:
          description: "Indicates the numbers of family, friends and/or cate team\
            \ persons helping the patient through their journey."
          maximum: 100
          minimum: 0
          title: supportPersonsNumber
          type: number
        drugAllergies:
          description: Known drug allergies such as concomitant medications.
          items:
            maxLength: 255
            minLength: 1
            type: string
          maxItems: 50
          minItems: 1
          title: drugAllergies
          type: array
          uniqueItems: true
        prescription:
          $ref: '#/components/schemas/Prescription'
      title: Clinical Details
    ServiceSubscription:
      anyOf:
      - $ref: '#/components/schemas/ServiceSubscriptionBasic'
      - $ref: '#/components/schemas/CaseManagementService'
      - $ref: '#/components/schemas/SupplyService'
      - $ref: '#/components/schemas/CopayService'
      - $ref: '#/components/schemas/PatientEducationService'
      - $ref: '#/components/schemas/RxBridgeService'
      description: "Main categories of services offered by BI directly or through\
        \ its partners to which a patient can sign up to.  \n\nEach service (e..g\
        \ Case Management) could provide additional \"service offerings\" not detailed\
        \ here. "
      discriminator:
        mapping:
          caseManagement: '#/components/schemas/CaseManagementService'
          patientEducation: '#/components/schemas/PatientEducationService'
          copay: '#/components/schemas/CopayService'
          rxBridge: '#/components/schemas/RxBridgeService'
          demoDeviceAutoInjection: '#/components/schemas/ServiceSubscriptionBasic'
          demoDevicePrefilledSyringe: '#/components/schemas/ServiceSubscriptionBasic'
          sharpsContainer: '#/components/schemas/SupplyService'
          travelColdCase: '#/components/schemas/SupplyService'
          printedEducationMaterials: '#/components/schemas/SupplyService'
        propertyName: serviceType
      title: Service Subscription
...
    PatientSupportDetails:
      description: Brand specific services enrollment and patient's health related
        data
      properties:
        clinicalDetails:
          $ref: '#/components/schemas/ClinicalDetails'
        insuranceTypes:
          description: "Indicates one or more medical insurance plans are available\
            \ and the plans types.\r\nFor example:\r\nCommercial\r\nHealth Exchange\r\
            \nManaged Medicaid\r\nMedicare\r\nState Medicaid\r\nTRICARE\r\nVA\r\n\
            Other Government\r\n"
          items:
            example: commertial
            maxLength: 255
            minLength: 1
            type: string
          maxItems: 10
          minItems: 1
          title: insuranceTypes
          type: array
          uniqueItems: true
        insurancePlans:
          description: Patient's list of pharmacy insurance plans.
          items:
            $ref: '#/components/schemas/InsurancePlan'
          title: insurancePlans
          type: array
          uniqueItems: true
        serviceSubscriptions:
          description: "List of services selected by the applicant,  and offered by\
            \ BI for the brand and region, that will help the patient throughout their\
            \ treatment journey.\r\n\r\nService availables are restricted to Copay\
            \ using this method. "
          items:
            $ref: '#/components/schemas/ServiceSubscription'
          maxItems: 100
          minItems: 1
          title: serviceSubscriptions
          type: array
      required:
      - clinicalDetails
      - serviceSubscriptions
      title: Patient Support Details
      x-examples: {}
...
    patch_patient_consents_brand_request:
      allOf:
      - properties:
          consentsAndPreferences:
            description: List of opt-ins and opt-outs the patient has provided to
              BI and partners
            items:
              $ref: '#/components/schemas/Consent'
            minItems: 1
            type: array
            uniqueItems: true
        required:
        - consentsAndPreferences
      - $ref: '#/components/schemas/DataProviderInfo'
      title: patch_patient_consents_brand_request
    patch_patient_consents_brand_200_response:
      example:
        consentsAndPreferences:
        - preferences:
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          name: Patient Support Program Communications
          status: optOut
        - preferences:
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          name: Patient Support Program Communications
          status: optOut
        - preferences:
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          name: Patient Support Program Communications
          status: optOut
        - preferences:
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          name: Patient Support Program Communications
          status: optOut
        - preferences:
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          - name: bestContactTime
            selectedOptions:
            - morning
            - morning
            - morning
            - morning
            - morning
          name: Patient Support Program Communications
          status: optOut
      properties:
        consentsAndPreferences:
          items:
            $ref: '#/components/schemas/Consent'
          maxItems: 100
          minItems: 1
          title: consentsAndPreferences
          type: array
      title: patch_patient_consents_brand_200_response
...
    patch_patients_id_services_brand_request:
      allOf:
      - $ref: '#/components/schemas/PatientSupportDetails'
      - $ref: '#/components/schemas/DataProviderInfo'
      title: patch_patients_id_services_brand_request
...
  securitySchemes:
    oauth2DEV:
      flows:
        clientCredentials:
          refreshUrl: ""
          scopes: {}
          tokenUrl: https://api-gw-dev.boehringer-ingelheim.com:8443/api/oauth/token
      type: oauth2
    oauth2QA:
      flows:
        clientCredentials:
          refreshUrl: ""
          scopes: {}
          tokenUrl: https://api-gw-qa.boehringer-ingelheim.com/api/oauth/token
      type: oauth2
    oauth2PROD:
      flows:
        clientCredentials:
          refreshUrl: ""
          scopes: {}
          tokenUrl: https://api-gw.boehringer-ingelheim.com/api/oauth/token
      type: oauth2